Update server.c
Added a logging function for the server, where we take the given phrase out of the log buffer with a log mutex lock and wait to consume from the log buffer, while the log bugger is empty. Then we remove the given words out of the log buffer and write it to the file, repeating for every needed phrase (and reseting the values as such).