Update server.c
Establishing error handling before server starts, checking for all variable initializations, if there is a problem, we write to stdout an error. If no problem, we run an infinite loop for incoming connections to be accepted by a socket descriptor, where there is another one created to deal with a connection, if one is successfully present.